Move App out of src root and into components/App.
